refactor quest file, add draul leptoc dialogue
